refactored code for navigation
