+++ app/src/main/java/com/entertainment/utils/navigationutils/SealedState.kt
@@ -1,0 +1,3 @@
+package com.entertainment.utils.navigationutils
+
+interface SealedState
